fix error states and clean up
