+++ Projects/Forever/Sources/Forever.swift
@@ -15,43 +15,45 @@ public struct ForeverView: View {
     public init() {}
 
     public var body: some View {
-        ScrollViewReader { value in
-            hForm {
-                VStack(spacing: 16) {
-                    HeaderView { scrollTo = 2 }.id(0)
-                    DiscountCodeSectionView().id(1)
-                    InvitationTable().id(2)
+        LoadingViewWithContent(ForeverStore.self, [.fetchForeverData], [.fetch]) {
+            ScrollViewReader { value in
+                hForm {
+                    VStack(spacing: 16) {
+                        HeaderView { scrollTo = 2 }.id(0)
+                        DiscountCodeSectionView().id(1)
+                        InvitationTable().id(2)
+                    }
+                }
+                .onChange(of: scrollTo) { newValue in
+                    if newValue != 0 {
+                        withAnimation {
+                            value.scrollTo(newValue, anchor: .top)
+                        }
+                        scrollTo = 0
+                    }
                 }
             }
-            .onChange(of: scrollTo) { newValue in
-                if newValue != 0 {
-                    withAnimation {
-                        value.scrollTo(newValue, anchor: .top)
-                    }
-                    scrollTo = 0
-                }
+            .onAppear {
+                store.send(.fetch)
             }
-        }
-        .onAppear {
-            store.send(.fetch)
-        }
-        .navigationBarItems(
-            trailing:
-                PresentableStoreLens(
-                    ForeverStore.self,
-                    getter: { state in
-                        state.foreverData?.monthlyDiscountPerReferral ?? .init(amount: 10, currency: "sek")
-                    }
-                ) { discountAmount in
+            .navigationBarItems(
+                trailing:
+                    PresentableStoreLens(
+                        ForeverStore.self,
+                        getter: { state in
+                            state.foreverData?.monthlyDiscountPerReferral ?? .init(amount: 10, currency: "sek")
+                        }
+                    ) { discountAmount in
                         Button(action: {
                             store.send(.showInfoSheet(discount: discountAmount.formattedAmount))
                         }) {
                             Image(uiImage: hCoreUIAssets.infoIcon.image)
                                 .foregroundColor(hTextColorNew.primary)
                         }
-                }
-        )
-        .trackOnAppear(hAnalyticsEvent.screenView(screen: .forever))
+                    }
+            )
+            .trackOnAppear(hAnalyticsEvent.screenView(screen: .forever))
+        }
     }
 }
