+++ Projects/Forever/Sources/Store/ForeverState.swift
@@ -1,0 +1,16 @@
+import Foundation
+import Presentation
+
+public struct ForeverState: StateProtocol {
+    public var hasSeenFebruaryCampaign: Bool {
+        didSet {
+            UserDefaults.standard.synchronize()
+        }
+    }
+    
+    public init() {
+        self.hasSeenFebruaryCampaign = false
+    }
+    
+    public var foreverData: ForeverData? = nil
+}
